Fixed Issue #43 BulkInsert can receive IDs out of order Updated nuget and assembly version to v1.4.6

+++ N.EntityFramework.Extensions/Data/BulkInsertResult.cs
@@ -6,6 +6,6 @@ namespace N.EntityFramework.Extensions
     internal class BulkInsertResult<T>
     {
         internal int RowsAffected { get; set; }
-        internal Dictionary<int, T> EntityMap { get; set; }
+        internal Dictionary<long, T> EntityMap { get; set; }
     }
 }
